requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
supabase
python-jose
jinja2
python-multipart
psycopg2-binary
sendgrid
twilio
